+++ utils/codegen/codegen-py/asn1ToRosMsg.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 # ==============================================================================
 # MIT License
@@ -30,6 +30,7 @@ import os
 from typing import Dict, List
 
 import jinja2
+from tqdm import tqdm
 
 from asn1CodeGenerationUtils import *
 
@@ -42,12 +43,12 @@ def parseCli():
     """
 
     parser = argparse.ArgumentParser(
-        description="Creates ROS .msg files from ASN1 definitions.",
+        description="Creates ROS .msg files from ASN.1 definitions.",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 
-    parser.add_argument("files", type=str, nargs="+", help="ASN1 files")
+    parser.add_argument("files", type=str, nargs="+", help="ASN.1 files")
     parser.add_argument("-o", "--output-dir", type=str, required=True, help="output directory")
-    parser.add_argument("-t", "--type", type=str, required=True, help="ASN1 type")
+    parser.add_argument("-t", "--type", type=str, required=True, help="ASN.1 type")
 
     args = parser.parse_args()
 
@@ -68,14 +69,16 @@ def loadJinjaTemplate() -> jinja2.environment.Template:
     return jinja_template
 
 
-def asn1TypeToRosMsg(type_name: str, asn1_type: Dict, asn1_types: Dict[str, Dict], asn1_values: Dict[str, Dict], asn1_raw: Dict[str, str], jinja_template: jinja2.environment.Template) -> str:
-    """Converts parsed ASN1 type information to a ROS message file string.
+def asn1TypeToRosMsg(type_name: str, asn1_type: Dict, asn1_types: Dict[str, Dict], asn1_values: Dict[str, Dict], asn1_sets: Dict[str, Dict], asn1_classes: Dict[str, Dict], asn1_raw: Dict[str, str], jinja_template: jinja2.environment.Template) -> str:
+    """Converts parsed ASN.1 type information to a ROS message file string.
 
     Args:
         type_name (str): type name
         asn1_type (Dict): type information
         asn1_types (Dict[str, Dict]): type information of all types by type
         asn1_values (Dict[str, Dict]): value information of all values by name
+        asn1_sets (Dict[str, Dict]): set information of all sets by name
+        asn1_classes (Dict[str, Dict]): class information of all classes by name
         asn1_raw (Dict[str, str]): raw string definition by type
         jinja_template (jinja2.environment.Template): jinja template
 
@@ -84,11 +87,11 @@ def asn1TypeToRosMsg(type_name: str, asn1_type: Dict, asn1_types: Dict[str, Dict
     """
 
     # build jinja context based on asn1 type information
-    jinja_context = asn1TypeToJinjaContext(type_name, asn1_type, asn1_types, asn1_values)
+    jinja_context = asn1TypeToJinjaContext(type_name, asn1_type, asn1_types, asn1_values, asn1_sets, asn1_classes)
     if jinja_context is None:
         return None
 
-    # add raw asn1 definition as comment to ROS .msg
+    # add raw asn1 definition as comment
     if type_name in asn1_raw:
         jinja_context["asn1_definition"] = asn1_raw[type_name].rstrip("\n")
 
@@ -134,10 +137,10 @@ def findDependenciesOfRosMessageType(parent_file_path: str, file_list: List[str]
                 if msg_type not in new_file_list and os.path.isfile(f"{os.path.dirname(parent_file_path)}/{msg_type}.msg"):
                     new_file_list.append(msg_type)
                     new_file_list = findDependenciesOfRosMessageType(f"{os.path.dirname(parent_file_path)}/{msg_type}.msg", new_file_list)
-    
+
     # make sure there are no duplicates and sort alphabetically
     new_file_list = sorted(list(set(new_file_list)))
-    
+
     return new_file_list
 
 def generateCMakeLists(msg_files: list, file_path: str, type: str) -> None:
@@ -200,25 +203,25 @@ def main():
 
     args = parseCli()
 
+    # parse ASN.1 files
+    print("Parsing ASN.1 files ...")
     asn1_docs, asn1_raw = parseAsn1Files(args.files)
-
     asn1_types = extractAsn1TypesFromDocs(asn1_docs)
     asn1_values = extractAsn1ValuesFromDocs(asn1_docs)
+    asn1_sets = extractAsn1SetsFromDocs(asn1_docs)
+    asn1_classes = extractAsn1ClassesFromDocs(asn1_docs)
 
     checkTypeMembersInAsn1(asn1_types)
 
+    # generate ROS .msg files
     jinja_template = loadJinjaTemplate()
-
-    for type_name, asn1_type in asn1_types.items():
-
-        ros_msg = asn1TypeToRosMsg(type_name, asn1_type, asn1_types, asn1_values, asn1_raw, jinja_template)
-
+    for type_name, asn1_type in (pbar := tqdm(asn1_types.items(), desc="Generating ROS .msg files")):
+        pbar.set_postfix_str(type_name)
+        ros_msg = asn1TypeToRosMsg(type_name, asn1_type, asn1_types, asn1_values, asn1_sets, asn1_classes, asn1_raw, jinja_template)
         exportRosMsg(ros_msg, type_name, args.output_dir)
 
-    # generate CMakelists.txt and remove all msg files that are not required
+    # generate CMakeLists.txt and remove all files that are not required for top-level message type
     msg_type = args.type.upper()
-    
-    # handle special cases
     if args.type == "cpm_ts":
         msg_type = "CollectivePerceptionMessage"
     elif args.type == "cam_ts":
@@ -229,13 +232,15 @@ def main():
         msg_type = "SPATEM"
     elif args.type == "vam_ts":
         msg_type = "VAM"
-
     msg_files = findDependenciesOfRosMessageType(os.path.join(args.output_dir, f"{msg_type}.msg"), [msg_type])
+    print("Generating CMakeLists.txt ...")
     generateCMakeLists(msg_files, os.path.join(args.output_dir, "../CMakeLists.txt"), args.type)
-    
+    print("Removing files not required for top-level message type ...")
     for f in glob.glob(os.path.join(args.output_dir, "*.msg")):
         if os.path.splitext(os.path.basename(f))[0] not in msg_files:
             os.remove(f)
+
+    print(f"Generated {len(msg_files)} ROS .msg files for {msg_type}")
 
 if __name__ == "__main__":