+++ etsi_its_conversion/etsi_its_conversion/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package etsi_its_conversion
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.0 (2024-12-10)
+------------------
+* Merge pull request `#51 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/51>`_ from ika-rwth-aachen/improvement/converter-launch-file
+  Improve converter launch file for better configuration
+* Merge pull request `#28 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/28>`_ from ika-rwth-aachen/feature/spatem-mapem
+  SPATEM/MAPEM Support
+* Contributors: Jean-Pierre Busch, Lennart Reiher
+
 2.4.0 (2024-11-19)
 ------------------
 * Merge branch 'main' into feature/denm-launchfile-transformation