+++ etsi_its_conversion/etsi_its_primitives_conversion/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package etsi_its_primitives_conversion
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.0 (2024-12-10)
+------------------
+
 2.4.0 (2024-11-19)
 ------------------
 * Merge branch 'main' into feature/denm-launchfile-transformation