+++ etsi_its_msgs/etsi_its_cam_msgs/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package etsi_its_cam_msgs
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.0 (2024-12-10)
+------------------
+* Merge pull request `#46 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/46>`_ from ika-rwth-aachen/improvement/formatting
+  Usability and formatting improvements for codegen-py
+* Merge pull request `#28 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/28>`_ from ika-rwth-aachen/feature/spatem-mapem
+  SPATEM/MAPEM Support
+* Contributors: Jean-Pierre Busch, Lennart Reiher
+
 2.4.0 (2024-11-19)
 ------------------
 * Merge pull request `#42 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/42>`_ from ika-rwth-aachen/feature/auto-cmake-generation