+++ etsi_its_msgs_utils/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package etsi_its_msgs_utils
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.0 (2024-12-10)
+------------------
+* Merge pull request `#48 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/48>`_ from ika-rwth-aachen/improvement/cpm-codegen-py
+  Enable CLASS and SET support in codegen-py
+* Merge pull request `#28 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/28>`_ from ika-rwth-aachen/feature/spatem-mapem
+  SPATEM/MAPEM Support
+* Contributors: Jean-Pierre Busch, Lennart Reiher
+
 2.4.0 (2024-11-19)
 ------------------
 * Merge pull request `#42 <https://github.com/ika-rwth-aachen/etsi_its_messages/issues/42>`_ from ika-rwth-aachen/feature/auto-cmake-generation